Flipped heading output to match compass (points in direction of field)

+++ examples/Module_Sketches/Module_10/Sensor_Tester/Sensor_Tester.ino
@@ -22,7 +22,7 @@ int capSample;
 // Output Formatting Variables
 String printOut;
 String sep = ", ";
-int timeStamp;
+long timeStamp;
 
 void setup() {
     // Start up Serial (wired) Communications
